texk/dvipdfmx, XeTeX: Use mfgets() parameters as for fgets()
git-svn-id: svn://tug.org/texlive/trunk/Build/source@36644 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/dvipdfm-x/mfileio.c
@@ -126,7 +126,7 @@ off_t xfile_size (FILE *file, const char *name)
 }
 
 /* Unlike fgets, mfgets works with \r, \n, or \r\n end of lines. */
-char *mfgets (char *buffer, unsigned long length, FILE *file) 
+char *mfgets (char *buffer, int length, FILE *file) 
 {
   int ch = 0, i = 0;
   while (i < length-1 && (ch = fgetc (file)) >= 0 && ch != '\n' && ch != '\r')
